Finish the uninstaller

+++ uninstaller.py
@@ -2,16 +2,27 @@ import os
 
 await ctx.send("Uninstalling DexScript...")
 
-os.remove("ballsdex/core/dexscript.py")
+if os.path.isfile("ballsdex/core/dexscript.py"):
+  os.remove("ballsdex/core/dexscript.py")
+
+exclude = [
+  "from ballsdex.core.dexscript import DexScript",
+  "        await self.add_cog(DexScript(self))"
+]
 
 # Add the ability to load the DexScript cog to the bot.py file.
 with open("ballsdex/core/bot.py", "r") as opened_file_1:
-  read = opened_file_1.read()
+  lines = opened_file_1.readlines()
+  content = ""
 
-  read.replace("from ballsdex.core.dexscript import DexScript\n", "")
-  read.replace("        await self.add_cog(DexScript(self))\n", "")
+  for index, line in enumerate(lines):
+    if line.rstrip() in exclude:
+      print("I HEARD SKIPPING")
+      continue
+
+    content += line
 
   with open("ballsdex/core/bot.py", "w") as opened_file_2:
-    opened_file_2.write(read)
+    opened_file_2.write(content)
 
 await ctx.send("DexScript has been uninstalled.\nRestart your bot for DexScript to be removed.")
